Initial version of the Quasar family of mod packs, involving a common core, a main version, a skyblock, and a survival variant.
All can be built from this, with .cfg files specifying the groups to use for a variant, and .csv files for all the mods for that group. Associated directories with the same name as the .cfg files (without the extension) contain the overrides folder with configs and such.

The information, credits, and licensing/mod pack permission data for each group of mods is in the respective group directory as MODS.txt, and are concatenated to make a pack variant's own MODS.txt.